backend/requirements.txt
Flask==3.1.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyJWT==2.6.0
bcrypt==4.0.1
cryptography==40.0.1
pymongo==4.7.2
scikit-learn
pandas
joblib